bye bye polynomial term operations\!
