+++ lib/collectAndCombine/CombineChecks.js
@@ -5,6 +5,14 @@ const PolynomialTermNode = require('../PolynomialTermNode');
 
 const CombineChecks = {};
 
+// Returns true if the node is an operation node with parameters that are
+// polynomial terms that can be combined in some way.
+CombineChecks.canSimplifyPolynomialTerms = function(node) {
+  return (CombineChecks.canAddLikeTermPolynomialNodes(node) ||
+          CombineChecks.canMultiplyLikeTermPolynomialNodes(node) ||
+          CombineChecks.canRearrangeCoefficient(node));
+};
+
 // Returns true if the nodes are symbolic terms with the same symbol and no
 // coefficients.
 CombineChecks.canMultiplyLikeTermPolynomialNodes = function(node) {
@@ -58,5 +66,28 @@ CombineChecks.canAddLikeTermPolynomialNodes = function(node) {
     return haveSameSymbol && haveSameExponent;
   });
 };
+
+// Returns true if the expression is a multiplication between a constant
+// and polynomial without a coefficient.
+CombineChecks.canRearrangeCoefficient = function(node) {
+  // implicit multiplication doesn't count as multiplication here, since it
+  // represents a single term.
+  if (node.op !== '*' || node.implicit) {
+    return false;
+  }
+  if (node.args.length !== 2) {
+    return false;
+  }
+  if (!NodeType.isConstantOrConstantFraction(node.args[1])) {
+    return false;
+  }
+  if (!PolynomialTermNode.isPolynomialTerm(node.args[0])) {
+    return false;
+  }
+
+  const polyNode = new PolynomialTermNode(node.args[0]);
+  return !polyNode.hasCoeff();
+};
+
 
 module.exports = CombineChecks;
