+++ lib/simplifyBasics/rearrangeCoefficient.js
@@ -5,13 +5,13 @@ const MathChangeTypes = require('../MathChangeTypes');
 const NodeCreator = require('../util/NodeCreator');
 const NodeStatus = require('../util/NodeStatus');
 const PolynomialTermNode = require('../PolynomialTermNode');
-const PolynomialTermOperations = require('../PolynomialTermOperations');
+const CombineChecks = require('../collectAndCombine/CombineChecks');
 
 // Rearranges something of the form x * 5 to be 5x, ie putting the coefficient
 // in the right place.
 // Returns a NodeStatus object
 function rearrangeCoefficient(node) {
-  if (!PolynomialTermOperations.canRearrangeCoefficient(node)) {
+  if (!CombineChecks.canRearrangeCoefficient(node)) {
     return NodeStatus.noChange(node);
   }
