+++ lib/util/removeUnnecessaryParens.js
@@ -2,7 +2,7 @@
 
 const LikeTermCollector = require('../collectAndCombine/LikeTermCollector');
 const NodeType = require('./NodeType');
-const PolynomialTermOperations = require('../PolynomialTermOperations');
+const CombineChecks = require('../collectAndCombine/CombineChecks');
 const PolynomialTermNode = require('../PolynomialTermNode');
 const resolvesToConstant = require('./resolvesToConstant');
 
@@ -159,7 +159,7 @@ function removeUnnecessaryParensInParenthesisNode(node) {
 function canCollectOrCombine(node) {
   return LikeTermCollector.canCollectLikeTerms(node) ||
     resolvesToConstant(node) ||
-    PolynomialTermOperations.canSimplifyPolynomialTerms(node);
+    CombineChecks.canSimplifyPolynomialTerms(node);
 }
 
 module.exports = removeUnnecessaryParens;
